% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-deg.R
\name{convert_deg}
\alias{convert_deg}
\title{Convert form of latitude and longitude from DMS to DEG}
\usage{
convert_deg(dms)
}
\arguments{
\item{dms}{Latitude or longitude in DMS format.}
}
\value{
Latitude or longitude converted to DEG format.
}
\description{
Converts latitude and longitude expressed in
DMS (Degree, Minute, Second) format to decimal notation,
DEG (Degree) format.
}
\examples{
convert_deg("40.3253820") # 40.54828
convert_deg("140.3554682") # 140.5985
}
