% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-traffic-accidents.R
\name{download_traffic_accidents_main}
\alias{download_traffic_accidents_main}
\alias{download_traffic_accidents_sub}
\alias{download_traffic_accidents_highway}
\alias{download_traffic_accidents_code}
\title{Download traffic accident data}
\usage{
download_traffic_accidents_main(destdir = NULL, year = 2021)

download_traffic_accidents_sub(destdir = NULL, year = 2021)

download_traffic_accidents_highway(destdir = NULL, year = 2021)

download_traffic_accidents_code(destdir = NULL, year = 2021)
}
\arguments{
\item{destdir}{A directory where downloaded file is to saved.}

\item{year}{Year the traffic accident data was recorded.}
}
\value{
A path string of the downloaded file
}
\description{
Download a file of traffic accident data from the National Police Agency's
Web page (https://www.npa.go.jp/publications/statistics/koutsuu/opendata/index_opendata.html).
The type of traffic accident data include main (honhyo) , sub (hojuhyo),
highway (kosokuhyo), or code (codebook).
The downloaded file will be saved in the specified directory
with the naming format \code{"traffic-accident-{year}-{type}.{ext}"}.
\code{ext} is the file extension: \code{"pdf"} if \code{type} is \code{"code"},
\code{"csv"} other \code{type}s.
}
\examples{
\dontrun{
# Download data sets.
download_traffic_accidents_main("data")
download_traffic_accidents_sub("data")
download_traffic_accidents_highway("data")
download_traffic_accidents_code("data")
}
}
