% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-traffic-accidents.R
\name{read_traffic_accidents_main}
\alias{read_traffic_accidents_main}
\alias{read_traffic_accidents_sub}
\alias{read_traffic_accidents_highway}
\alias{read_traffic_accidents_code}
\title{Read traffic accident data}
\usage{
read_traffic_accidents_main(file)

read_traffic_accidents_sub(file)

read_traffic_accidents_highway(file)

read_traffic_accidents_code(file)
}
\arguments{
\item{file}{A file path of data to read}
}
\value{
Data read in \code{tibble} format
}
\description{
Read a file of traffic accident data for a specified path.
The type of traffic accident data include main (honhyo) , sub (hojuhyo),
highway (kosokuhyo), or code (codebook).
}
\examples{
\dontrun{
# Read traffic accident main data(honhyo)
file <- download_traffic_accidents_main("data")
data <- read_traffic_accidents_main(file)
}
}
