% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtatools-package.R
\docType{package}
\name{rtatools-package}
\alias{rtatools}
\alias{rtatools-package}
\title{rtatools: Functions for road traffic accident data analysis}
\description{
The `rtatools` is a package for road traffic accident data analysis. This package provides the functions to analyze accident data in Japan, such as downloading the traffic accident data and converting the coordinates of accident points.
}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
